Factor patches together
no-patch smoke test works
