--- src/methods-mutation.js
@@ -1,274 +1,0 @@
-/**
-@license
-Copyright (c) 2016 The Polymer Project Authors. All rights reserved.
-This code may only be used under the BSD style license found at http://polymer.github.io/LICENSE.txt
-The complete set of authors may be found at http://polymer.github.io/AUTHORS.txt
-The complete set of contributors may be found at http://polymer.github.io/CONTRIBUTORS.txt
-Code distributed by Google as part of the polymer project is also
-subject to an additional IP rights grant found at http://polymer.github.io/PATENTS.txt
-*/
-
-import * as utils from './utils.js';
-import * as logicalTree from './logical-tree.js';
-import * as nativeMethods from './native-methods.js';
-import {accessors as nativeAccessors} from './native-tree.js';
-import {getRootNode} from './methods-get-root-node.js';
-import {shadyDataForNode} from './shady-data.js';
-import {ElementAccessors as accessors} from './accessors.js';
-import {currentScopeForNode, currentScopeIsCorrect, addShadyScoping, removeShadyScoping, replaceShadyScoping, treeVisitor} from './style-scoping.js';
-
-const {parentNode: nativeParentNode} = nativeAccessors;
-
-const doc = window.document;
-
-// Patched `insertBefore`. Note that all mutations that add nodes are routed
-// here. When a <slot> is added or a node is added to a host with a shadowRoot
-// with a slot, a standard dom `insert` call is aborted and `_asyncRender`
-// is called on the relevant shadowRoot. In all other cases, a standard dom
-// `insert` can be made, but the location and ref_node may need to be changed.
-/**
- * @param {!Node} parent
- * @param {Node} node
- * @param {Node=} ref_node
- */
-export function insertBefore(parent, node, ref_node) {
-  // optimization: assume native insertBefore is ok if the nodes are not in the document.
-  if (parent.ownerDocument !== doc && node.ownerDocument !== doc) {
-    nativeMethods.insertBefore.call(parent, node, ref_node);
-    return;
-  }
-  if (node === parent) {
-    throw Error(`Failed to execute 'appendChild' on 'Node': The new child element contains the parent.`);
-  }
-  if (ref_node) {
-    const refData = shadyDataForNode(ref_node);
-    const p = refData && refData.parentNode;
-    if ((p !== undefined && p !== parent) ||
-      (p === undefined && nativeParentNode(ref_node) !== parent)) {
-      throw Error(`Failed to execute 'insertBefore' on 'Node': The node ` +
-       `before which the new node is to be inserted is not a child of this node.`);
-    }
-  }
-  if (ref_node === node) {
-    return node;
-  }
-  /** @type {!Array<!HTMLSlotElement>} */
-  let slotsAdded = [];
-  /** @type {function(!Node, string): void} */
-  let scopingFn = addShadyScoping;
-  let ownerRoot = utils.ownerShadyRootForNode(parent);
-  /** @type {string} */
-  let newScopeName;
-  if (ownerRoot) {
-    // get scope name from new ShadyRoot host
-    newScopeName = ownerRoot.host.localName;
-  } else {
-    // borrow parent's scope name
-    newScopeName = currentScopeForNode(parent);
-  }
-  // remove from existing location
-  const parentNode = accessors.parentNode.get.call(node);
-  if (parentNode) {
-    // NOTE: avoid node.removeChild as this *can* trigger another patched
-    // method (e.g. custom elements) and we want only the shady method to run.
-    // The following table describes what style scoping actions should happen as a result of this insertion.
-    // document -> shadowRoot: replace
-    // shadowRoot -> shadowRoot: replace
-    // shadowRoot -> shadowRoot of same type: do nothing
-    // shadowRoot -> document: allow unscoping
-    // document -> document: do nothing
-    // The "same type of shadowRoot" and "document to document cases rely on `currentScopeIsCorrect` returning true
-    const oldScopeName = currentScopeForNode(node);
-    removeChild(parentNode, node, Boolean(ownerRoot) || !(getRootNode(node) instanceof ShadowRoot));
-    scopingFn = (node, newScopeName) => {
-      replaceShadyScoping(node, newScopeName, oldScopeName);
-    };
-  }
-  // add to new parent
-  let allowNativeInsert = true;
-  // TODO(sorvell): svg hack is just for testing.
-  const needsScoping = node['__noInsertionPoint'] === undefined && node.localName !== 'svg' && !currentScopeIsCorrect(node, newScopeName);
-  if (ownerRoot) {
-    // in a shadowroot, only tree walk if new insertion points may have been added, or scoping is needed
-    if (!node['__noInsertionPoint'] || needsScoping) {
-      treeVisitor(node, (node) => {
-        if (node.localName === 'slot') {
-          slotsAdded.push(/** @type {!HTMLSlotElement} */(node));
-        }
-        if (needsScoping) {
-          scopingFn(node, newScopeName);
-        }
-      });
-    }
-  } else if (needsScoping) {
-    // in a document or disconnected tree, replace scoping if necessary
-    const oldScopeName = currentScopeForNode(node);
-    treeVisitor(node, (node) => {
-      replaceShadyScoping(node, newScopeName, oldScopeName);
-    });
-  }
-  if (slotsAdded.length) {
-    ownerRoot._addSlots(slotsAdded);
-  }
-  // if a slot is added, must render containing root.
-  if (parent.localName === 'slot' || slotsAdded.length) {
-    if (ownerRoot) {
-      ownerRoot._asyncRender();
-    }
-  }
-  if (utils.isTrackingLogicalChildNodes(parent)) {
-    logicalTree.recordInsertBefore(node, parent, ref_node);
-    // when inserting into a host with a shadowRoot with slot, use
-    // `shadowRoot._asyncRender()` via `attach-shadow` module
-    const parentData = shadyDataForNode(parent);
-    if (utils.hasShadowRootWithSlot(parent)) {
-      parentData.root._asyncRender();
-      allowNativeInsert = false;
-    // when inserting into a host with shadowRoot with NO slot, do nothing
-    // as the node should not be added to composed dome anywhere.
-    } else if (parentData.root) {
-      allowNativeInsert = false;
-    }
-  }
-  if (allowNativeInsert) {
-    // if adding to a shadyRoot, add to host instead
-    let container = utils.isShadyRoot(parent) ?
-      /** @type {ShadowRoot} */(parent).host : parent;
-    // if ref_node, get the ref_node that's actually in composed dom.
-    if (ref_node) {
-      ref_node = firstComposedNode(ref_node);
-      nativeMethods.insertBefore.call(container, node, ref_node);
-    } else {
-      nativeMethods.appendChild.call(container, node);
-    }
-  // Since ownerDocument is not patched, it can be incorrect afer this call
-  // if the node is physically appended via distribution. This can result
-  // in the custom elements polyfill not upgrading the node if it's in an inert doc.
-  // We correct this by calling `adoptNode`.
-  } else if (node.ownerDocument !== parent.ownerDocument) {
-    parent.ownerDocument.adoptNode(node);
-  }
-  scheduleObserver(parent, node);
-  return node;
-}
-
-/**
- * Patched `removeChild`. Note that all dom "removals" are routed here.
- * Removes the given `node` from the element's `children`.
- * This method also performs dom composition.
- * @param {Node} parent
- * @param {Node} node
- * @param {boolean=} skipUnscoping
-*/
-export function removeChild(parent, node, skipUnscoping = false) {
-  if (parent.ownerDocument !== doc) {
-    return nativeMethods.removeChild.call(parent, node);
-  }
-  if (accessors.parentNode.get.call(node) !== parent) {
-    throw Error('The node to be removed is not a child of this node: ' +
-      node);
-  }
-  let preventNativeRemove;
-  let ownerRoot = utils.ownerShadyRootForNode(node);
-  let removingInsertionPoint;
-  const parentData = shadyDataForNode(parent);
-  if (utils.isTrackingLogicalChildNodes(parent)) {
-    logicalTree.recordRemoveChild(node, parent);
-    if (utils.hasShadowRootWithSlot(parent)) {
-      parentData.root._asyncRender();
-      preventNativeRemove = true;
-    }
-  }
-  // unscope a node leaving a ShadowRoot if ShadyCSS is present, and this node
-  // is not going to be rescoped in `insertBefore`
-  if (utils.getScopingShim() && !skipUnscoping && ownerRoot) {
-    const oldScopeName = currentScopeForNode(node);
-    treeVisitor(node, (node) => {
-      removeShadyScoping(node, oldScopeName);
-    });
-  }
-  removeOwnerShadyRoot(node);
-  // if removing slot, must render containing root
-  if (ownerRoot) {
-    let changeSlotContent = parent && parent.localName === 'slot';
-    if (changeSlotContent) {
-      preventNativeRemove = true;
-    }
-    removingInsertionPoint = ownerRoot._removeContainedSlots(node);
-    if (removingInsertionPoint || changeSlotContent) {
-      ownerRoot._asyncRender();
-    }
-  }
-  if (!preventNativeRemove) {
-    // if removing from a shadyRoot, remove from host instead
-    let container = utils.isShadyRoot(parent) ?
-      /** @type {ShadowRoot} */(parent).host :
-      parent;
-    // not guaranteed to physically be in container; e.g.
-    // (1) if parent has a shadyRoot, element may or may not at distributed
-    // location (could be undistributed)
-    // (2) if parent is a slot, element may not ben in composed dom
-    if (!(parentData.root || node.localName === 'slot') ||
-      (container === nativeParentNode(node))) {
-      nativeMethods.removeChild.call(container, node);
-    }
-  }
-  scheduleObserver(parent, null, node);
-  return node;
-}
-
-function removeOwnerShadyRoot(node) {
-  // optimization: only reset the tree if node is actually in a root
-  if (hasCachedOwnerRoot(node)) {
-    let c$ = accessors.childNodes.get.call(node);
-    for (let i=0, l=c$.length, n; (i<l) && (n=c$[i]); i++) {
-      removeOwnerShadyRoot(n);
-    }
-  }
-  const nodeData = shadyDataForNode(node);
-  if (nodeData) {
-    nodeData.ownerShadyRoot = undefined;
-  }
-}
-
-function hasCachedOwnerRoot(node) {
-  const nodeData = shadyDataForNode(node);
-  return Boolean(nodeData && nodeData.ownerShadyRoot !== undefined);
-}
-
-/**
- * Finds the first flattened node that is composed in the node's parent.
- * If the given node is a slot, then the first flattened node is returned
- * if it exists, otherwise advance to the node's nextSibling.
- * @param {Node} node within which to find first composed node
- * @returns {Node} first composed node
- */
-function firstComposedNode(node) {
-  let composed = node;
-  if (node && node.localName === 'slot') {
-    const nodeData = shadyDataForNode(node);
-    const flattened = nodeData && nodeData.flattenedNodes;
-    composed = flattened && flattened.length ? flattened[0] :
-      firstComposedNode(node.nextSibling);
-  }
-  return composed;
-}
-
-/**
- * @param {Node} node
- * @param {Node=} addedNode
- * @param {Node=} removedNode
- */
-function scheduleObserver(node, addedNode, removedNode) {
-  const nodeData = shadyDataForNode(node);
-  const observer = nodeData && nodeData.observer;
-  if (observer) {
-    if (addedNode) {
-      observer.addedNodes.push(addedNode);
-    }
-    if (removedNode) {
-      observer.removedNodes.push(removedNode);
-    }
-    observer.schedule();
-  }
-}
